backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-dotenv==1.0.1
google-cloud-firestore==2.16.0
google-generativeai==0.7.2
google-cloud-aiplatform==1.57.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.1
aiofiles==23.2.1
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
pyaudio==0.2.14
librosa==0.10.2
soundfile==0.12.1
scipy==1.13.1
tenacity==8.3.0
structlog==24.2.0
prometheus-fastapi-instrumentator==7.0.0
redis==5.0.4
celery==5.4.0
